+++ Makefile.am
@@ -132,7 +132,7 @@ src_libfabric_la_SOURCES =			\
 	include/fi_mem.h			\
 	include/fi_osd.h			\
 	include/fi_proto.h			\
-	include/fi_rbuf.h			\
+	include/ofi_rbuf.h			\
 	include/fi_shm.h			\
 	include/ofi_signal.h			\
 	include/fi_util.h			\
