+++ prov/gni/include/gnix.h
@@ -59,7 +59,7 @@
 #include <fi.h>
 #include <ofi_atomic.h>
 #include <fi_enosys.h>
-#include <fi_rbuf.h>
+#include <ofi_rbuf.h>
 #include <fi_list.h>
 #include <ofi_file.h>
