+++ prov/shm/src/smr.h
@@ -54,7 +54,7 @@
 #include <fi.h>
 #include <fi_enosys.h>
 #include <fi_shm.h>
-#include <fi_rbuf.h>
+#include <ofi_rbuf.h>
 #include <fi_list.h>
 #include <ofi_signal.h>
 #include <fi_util.h>
