+++ prov/udp/src/udpx.h
@@ -55,7 +55,7 @@
 
 #include <fi.h>
 #include <fi_enosys.h>
-#include <fi_rbuf.h>
+#include <ofi_rbuf.h>
 #include <fi_list.h>
 #include <ofi_signal.h>
 #include <fi_util.h>
